requirements.txt
pandas
pillow
pytorch-lightning
PyYAML
torch
torchvision
tqdm
wandb